+++ CHANGELOG.md
@@ -1,6 +1,9 @@
-## [3.0.1]
+## [3.1.0] BREAKING CHANGE
 
-* added `enabled` variable inside `FormKeyboardActions` in case you don't want to use `FormKeyboardActions` widget (tablets for example).
+* API improved
+* `KeyboardCustomInput` was added to help you to create custom keyboards in an easy way.
+* `FormKeyboardActions` was renamed to `KeyboardActions`.
+* added `enabled` variable inside `KeyboardActions` in case you don't want to use `KeyboardActions` widget (tablets for example).
 * Material color is transparent to avoid issues with the parent container.
 * Fixed issue with Stack and Scaffold.